+++ src/matterbridgeEndpoint.ts
@@ -1466,6 +1466,9 @@ export class MatterbridgeEndpoint extends Endpoint {
    * @param {DoorLock.LockState} [lockState=DoorLock.LockState.Locked] - The initial state of the lock (default: Locked).
    * @param {DoorLock.LockType} [lockType=DoorLock.LockType.DeadBolt] - The type of the lock (default: DeadBolt).
    * @returns {this} The current MatterbridgeEndpoint instance for chaining.
+   *
+   * @remarks
+   * All operating modes NOT supported by a lock SHALL be set to one. The value of the OperatingMode enumeration defines the related bit to be set.
    */
   createDefaultDoorLockClusterServer(lockState = DoorLock.LockState.Locked, lockType = DoorLock.LockType.DeadBolt) {
     this.behaviors.require(MatterbridgeDoorLockServer.enable({ events: { doorLockAlarm: true, lockOperation: true, lockOperationError: true } }), {
@@ -1473,7 +1476,7 @@ export class MatterbridgeEndpoint extends Endpoint {
       lockState,
       lockType,
       actuatorEnabled: false,
-      supportedOperatingModes: { normal: true, vacation: false, privacy: false, noRemoteLockUnlock: false, passage: false },
+      supportedOperatingModes: { normal: false, vacation: true, privacy: true, noRemoteLockUnlock: true, passage: true },
     });
     return this;
   }
@@ -1552,7 +1555,7 @@ export class MatterbridgeEndpoint extends Endpoint {
   createDefaultSmokeCOAlarmClusterServer(smokeState = SmokeCoAlarm.AlarmState.Normal, coState = SmokeCoAlarm.AlarmState.Normal) {
     this.behaviors.require(
       MatterbridgeSmokeCoAlarmServer.with(SmokeCoAlarm.Feature.SmokeAlarm, SmokeCoAlarm.Feature.CoAlarm).enable({
-        events: { smokeAlarm: true, interconnectSmokeAlarm: true, coAlarm: true, interconnectCoAlarm: true, lowBattery: true, hardwareFault: true, endOfService: true, selfTestComplete: true, alarmMuted: true, muteEnded: true, allClear: true },
+        events: { smokeAlarm: true, interconnectSmokeAlarm: false, coAlarm: true, interconnectCoAlarm: false, lowBattery: true, hardwareFault: true, endOfService: true, selfTestComplete: true, alarmMuted: true, muteEnded: true, allClear: true },
       }),
       {
         smokeState,
@@ -1563,8 +1566,56 @@ export class MatterbridgeEndpoint extends Endpoint {
         testInProgress: false,
         hardwareFaultAlert: false,
         endOfServiceAlert: SmokeCoAlarm.EndOfService.Normal,
-        interconnectSmokeAlarm: SmokeCoAlarm.AlarmState.Normal,
-        interconnectCoAlarm: SmokeCoAlarm.AlarmState.Normal,
+        // interconnectSmokeAlarm: SmokeCoAlarm.AlarmState.Normal,
+        // interconnectCoAlarm: SmokeCoAlarm.AlarmState.Normal,
+      },
+    );
+    return this;
+  }
+
+  /**
+   * Creates a smoke only SmokeCOAlarm Cluster Server.
+   *
+   * @param {SmokeCoAlarm.AlarmState} smokeState - The state of the smoke alarm. Defaults to SmokeCoAlarm.AlarmState.Normal.
+   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
+   */
+  createSmokeOnlySmokeCOAlarmClusterServer(smokeState = SmokeCoAlarm.AlarmState.Normal) {
+    this.behaviors.require(
+      MatterbridgeSmokeCoAlarmServer.with(SmokeCoAlarm.Feature.SmokeAlarm).enable({
+        events: { smokeAlarm: true, interconnectSmokeAlarm: false, coAlarm: false, interconnectCoAlarm: false, lowBattery: true, hardwareFault: true, endOfService: true, selfTestComplete: true, alarmMuted: true, muteEnded: true, allClear: true },
+      }),
+      {
+        smokeState,
+        expressedState: SmokeCoAlarm.ExpressedState.Normal,
+        batteryAlert: SmokeCoAlarm.AlarmState.Normal,
+        deviceMuted: SmokeCoAlarm.MuteState.NotMuted,
+        testInProgress: false,
+        hardwareFaultAlert: false,
+        endOfServiceAlert: SmokeCoAlarm.EndOfService.Normal,
+      },
+    );
+    return this;
+  }
+
+  /**
+   * Creates a co only SmokeCOAlarm Cluster Server.
+   *
+   * @param {SmokeCoAlarm.AlarmState} coState - The state of the CO alarm. Defaults to SmokeCoAlarm.AlarmState.Normal.
+   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
+   */
+  createCoOnlySmokeCOAlarmClusterServer(coState = SmokeCoAlarm.AlarmState.Normal) {
+    this.behaviors.require(
+      MatterbridgeSmokeCoAlarmServer.with(SmokeCoAlarm.Feature.CoAlarm).enable({
+        events: { smokeAlarm: false, interconnectSmokeAlarm: false, coAlarm: true, interconnectCoAlarm: false, lowBattery: true, hardwareFault: true, endOfService: true, selfTestComplete: true, alarmMuted: true, muteEnded: true, allClear: true },
+      }),
+      {
+        coState,
+        expressedState: SmokeCoAlarm.ExpressedState.Normal,
+        batteryAlert: SmokeCoAlarm.AlarmState.Normal,
+        deviceMuted: SmokeCoAlarm.MuteState.NotMuted,
+        testInProgress: false,
+        hardwareFaultAlert: false,
+        endOfServiceAlert: SmokeCoAlarm.EndOfService.Normal,
       },
     );
     return this;